sprint_11_add-friends-like исправление ошибок

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -9,7 +9,7 @@ public interface FilmStorage {
 
     void updateFilm(Film film);
 
-    void deleteFilm(Integer filmId);
+    void deleteFilm(Long filmId);
 
     Film getFilm(Long filmId);
